updated/fixed signalAnalysis

+++ HeavyChHiggsToTauNu/python/HChSignalAnalysisParameters_cff.py
@@ -138,7 +138,7 @@ MET = cms.untracked.PSet(
 bTagging = cms.untracked.PSet(
     discriminator = cms.untracked.string("trackCountingHighEffBJetTags"),
     discriminatorCut = cms.untracked.double(2.0),
-    ptCut = cms.untracked.double(20),
+    ptCut = cms.untracked.double(30),
     etaCut = cms.untracked.double(2.4),
     minNumber = cms.untracked.uint32(1)
 )
